Fixup highlight colors for errors and spelling
Some of the color combinations are not readable if the cursor is not on
the same line the error happened. Remove the background color from
errors and SpellBad so the text is still emphasized but can actually be
read.

Also disable SpellCap - it's usually complained about when a sentence
starts with a name which is supposed to be lowercase, or when the format
syntax uses a dot. I don't need the false positives colored in
white-on-lightblue combo (or any other).

Signed-off-by: Veronika Kabatova <veronicca114@gmail.com>